pyproject.toml
[project]
name = "serveease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdullah-1121", email = "muhmmad11abdullah21@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "firebase-admin>=6.9.0",
    "langsmith[openai-agents]>=0.4.1",
    "openai-agents>=0.0.15",
]

[project.scripts]
serveease = "serveease:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
